# crates/wire-schema/openapi.json
{
  "components": {
    "parameters": {
      "ProjectHeader": {
        "description": "Story 0.11 substrate (Phase 3 decision L2). Identifies the target project. Phase 2 single-project deployments accept the header for forward-compatibility: the value must equal the configured project name (case-insensitive after trim) or the reserved sentinel `default`. Mismatching values return 403 `project_not_found`. Absent header is accepted in Phase 2 (with a one-time per-process WARN log) and will become required in Phase 4 multi-project mode.",
        "in": "header",
        "name": "X-OpenGEO-Project",
        "required": false,
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Missing, malformed, or revoked X-OpenGEO-API-Key. The body shape is intentionally opaque so an unauthenticated caller cannot distinguish 'wrong key' from 'auth backend down'."
      }
    },
    "schemas": {
...
      "AuditRunItem": {
        "type": "object",
        "required": [
          "id",
          "target",
          "overall_score",
          "pages_crawled",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "target": {
            "type": "string"
          },
          "overall_score": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "pages_crawled": {
            "type": "integer",
            "minimum": 0
          },
          "gate_passed": {
            "type": "boolean",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AuditRunList": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditRunItem"
            }
          }
        }
      },
...
      "Error": {
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "type": "object"
      },
...
      "RunListResponse": {
        "properties": {
          "runs": {
            "items": {
              "type": "object"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "in": "header",
        "name": "X-OpenGEO-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Phase 2 v1 surface — read endpoints, prompt run submission, and the SSE event stream. Auth: X-OpenGEO-API-Key header carrying an `ogeo_<32 chars>` key. Story 0.11 (Phase 3 substrate) adds the optional X-OpenGEO-Project header — Phase 2 accepts it for forward compatibility; Phase 4 will require it. Story 15.1 adds three `/v1/setup/*` endpoints (deployment UX substrate).",
    "title": "OpenGEO Public REST API",
    "version": "0.3.1"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/v1/audit/runs": {
      "get": {
        "operationId": "listAuditRuns",
        "summary": "Roadmap Epic 32 — persisted site-audit history for the project, newest first.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProjectHeader"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditRunList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/v1/runs": {
      "get": {
        "operationId": "listRuns",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProjectHeader"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "maximum": 500,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunListResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "summary": "List recent Prompt Runs"
      }
    },
...
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "servers": [
    {
      "description": "Hosted production",
      "url": "https://api.opengeo.dev"
    },
    {
      "description": "Local Compose stack",
      "url": "http://127.0.0.1:8788"
    }
  ]
}